Preparatory work to add a silent response
Added a code enum to own our own HTTP code source so we can better work with Psalm annotations. We improved our code to the point that we are now at Psalm level 3, these code improvements have also had a positive effect on PHPStan as the last "phpstan-ignore" annotations could now be removed.